// ui/packages/user-app/src/api/models/pointHistoryItemResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Roulette API
 * 룰렛 게임 API 문서
 * OpenAPI spec version: 1.0.0
 */
import type { PointHistoryItemResponseType } from './pointHistoryItemResponseType';

export interface PointHistoryItemResponse {
  id: number;
  amount: number;
  type: PointHistoryItemResponseType;
  createdAt: string;
  expiresAt?: string;
}