// ui/packages/user-app/src/api/models/pointHistoryItemResponseType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Roulette API
 * 룰렛 게임 API 문서
 * OpenAPI spec version: 1.0.0
 */

export type PointHistoryItemResponseType =
  (typeof PointHistoryItemResponseType)[keyof typeof PointHistoryItemResponseType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PointHistoryItemResponseType = {
  EARN: 'EARN',
  USE: 'USE',
  REFUND: 'REFUND',
  RECLAIM: 'RECLAIM',
  EXPIRE: 'EXPIRE',
} as const;